Refactor tree incident and work item management
- Updated TreeIncidentDetailDto to use string for Status and initialized Images list.
- Implemented UpdateTreeCommand to handle tree updates with validation.
- Removed unused GetAllTreesQuery and GetTreeByIdQuery classes.
- Enhanced GetTreeMapQuery to fetch tree data with coordinates.
- Changed Latitude and Longitude types in TreeMapDto from decimal to double.
- Deleted obsolete GetTreesQuery class.
- Refactored AssignWorkCommand to use Work.Create method for work creation.
- Updated UpdateWorkItemCommand to utilize the new Update method in Work entity.
- Modified Plan entity to encapsulate creation and update logic.
- Enhanced TreeIncident entity to include reporter details and improved creation logic.
- Refactored Work entity to encapsulate creation and update logic.
- Removed unsupported ColourException and Colour value object.
- Updated ApplicationDbContextInitialiser to use new entity creation methods.
- Refactored MaintenanceJobService to use Work.Create for work assignments.
- Updated API endpoints for tree incidents and trees to include new commands and queries.
- Added new commands for creating incidents and sending feedback.
- Implemented DeleteTreeCommand for tree deletion functionality.
- Created GetTreeDetailQuery and GetWorkItemDetailQuery for detailed retrieval of tree and work item data.

+++ backend/src/Application/Trees/Queries/GetTreeMap/TreeMapVm.cs
@@ -11,6 +11,6 @@ public class TreeMapDto
     public string? Name { get; init; }
     public string? Condition { get; init; }
     public string? TreeTypeName { get; init; }
-    public decimal? Latitude { get; init; }
-    public decimal? Longitude { get; init; }
+    public double? Latitude { get; init; }
+    public double? Longitude { get; init; }
 }
